+++ integration-tests/src/test/scala/peloton/DurableStateStoreMySQLSpec.scala
@@ -1,0 +1,4 @@
+package peloton
+
+class DurableStateStoreMySQLSpec extends DurableStateStoreSpec:
+  val config = MySQLSpec.testContainerConfig