+++ integration-tests/src/test/scala/peloton/EventStoreMySQLSpec.scala
@@ -1,0 +1,4 @@
+package peloton
+
+class EventStoreMySQLSpec extends EventStoreSpec:
+  val config = MySQLSpec.testContainerConfig